requirements.txt
aiohttp==3.10.5
altair==5.3.0
faiss-cpu
google-search-results==2.4.2
langchain==0.2.15
langchain-community==0.2.15
langchain-experimental==0.0.64
langchain-openai==0.1.23
numexpr==2.8.4
numpy==1.26.4
openai==1.57.4
openpyxl
pandas==2.2.2
pathlib==1.0.1
protobuf==5.29.1
pydantic==2.7.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.0
pyyaml==6.0.1
requests==2.32.3
sqlalchemy==2.0.31
streamlit==1.40.2
tenacity==8.4.1
tiktoken==0.7.0
tqdm==4.66.4
typing-extensions==4.12.2
tzdata==2024.1
st-weaviate-connection==0.1.0
weaviate-client==4.9.6
dspy==2.5.42
